backend/requirements.txt
torch>=1.8.0
torchvision>=0.9.0
dassl>=0.1.0
scikit-learn>=1.0.2
numpy>=1.21.0
motor>=3.1.1
pydantic>=1.8.2
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0
Pillow>=8.0.0
openai>=1.0.0 